requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
python-multipart==0.0.9
ultralytics==8.1.2
opencv-python-headless==4.9.0.80
numpy==1.26.4
Pillow==10.2.0
loguru==0.7.2